requirements.txt
Django
djangorestframework
elasticsearch
elastic-app-search
sentence-transformers
social-auth-app-django
